+++ P_260729/Plugins/InventorySystem/Source/InventorySystem/Public/UI/InventoryDuckovSlotWidgetBase.h
@@ -23,6 +23,7 @@ protected:
 	virtual void NativeOnMouseLeave(const FPointerEvent& InMouseEvent) override;
 	virtual FReply NativeOnPreviewMouseButtonDown(const FGeometry& InGeometry, const FPointerEvent& InMouseEvent) override;
 	virtual FReply NativeOnMouseButtonDown(const FGeometry& InGeometry, const FPointerEvent& InMouseEvent) override;
+	virtual FReply NativeOnMouseButtonUp(const FGeometry& InGeometry, const FPointerEvent& InMouseEvent) override;
 	virtual FReply NativeOnMouseButtonDoubleClick(const FGeometry& InGeometry, const FPointerEvent& InMouseEvent) override;
 	virtual void NativeOnDragDetected(
 		const FGeometry& InGeometry,
@@ -56,4 +57,5 @@ protected:
 	TObjectPtr<UBorder> Border_QuantityBadge = nullptr;
 
 	bool bDragHovered = false;
+	bool bPendingQuickTransfer = false;
 };